+++ tower-service-util/Cargo.toml
@@ -26,3 +26,7 @@ either = { version = "1.5.1", optional = true }
 futures = "0.1.23"
 tower-service = "0.2.0"
 tokio-io = { version = "0.1.12", optional = true }
+
+[dev-dependencies]
+tower = { version = "0.1.0", path = "../tower" }
+tokio-mock-task = "0.1"
